+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.9.0] - 2019-01-26
+### Added
+- Pass more informations to `BuyButton`.
+
 ## [2.8.0] - 2019-01-23
 
 ## [2.7.2] - 2019-01-22